Changes related to new geth defaults
Geth now adds a gasPrice to dynamic fee transactions. The value for gasPrice that is set is equal to the maxFeePerGas value. Changes were made to modify_transaction since we build a new transaction based on values from the original, pending transaction. This transaction would now have all 3 fields and eth_sendTransaction would fail. We need to pop this value back out before replacing the transaction if it is equal to the expected value (maxFeePerGas).