bugfix: Unique caches for simple_cache_middleware instances
At present, if the ``simple_cache_middleware`` is imported and added to multiple instances of ``Web3``, these instances end up sharing a cache. This happens because the middleware is only constructed once and where we instantiate the cache if it is ``None`` is in the ``construct_simple_cache_middleware`` method. By moving the cache instantiation into the actual ``simple_cache_middleware``, we ensure that with each call to ``simple_cache_middleware``, a fresh cache will be instantiated if the ``cache`` was ``None`` to begin with.