Drop support for python 3.7:
We will have more peripheral updates to make due to the dropping of
python 3.7, such as forgoing some typing imports from typing_extensions,
etc. For now, we have PRs that are already expecting python 3.7 to be
dropped and all tests fail there. This commit aims to only drop the tests,
the CI support for py37, and the claim that the library supports 3.7.2 and above.